Moved sql_api_generate_jwt out of hooks folder
