+++ tests/providers/snowflake/utils/test_sql_api_generate_jwt.py
@@ -20,8 +20,7 @@ import pytest
 from cryptography.hazmat.backends import default_backend as crypto_default_backend
 from cryptography.hazmat.primitives import serialization as crypto_serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
-
-from airflow.providers.snowflake.hooks.sql_api_generate_jwt import JWTGenerator
+from providers.snowflake.utils.sql_api_generate_jwt import JWTGenerator
 
 _PASSWORD = "snowflake42"
